firmware: Restore IHEX-related conversion rules
Upstream commit 5620a0d ("firmware: delete in-kernel firmware")
removes the in-kernel firmware support, and commit
df85b2d ("firmware: Restore support for built-in firmware") added it
back, with minimal Makefile rules.

Samsung uses this support again, though it's been removed/chopped.
They added *.bin firmwares and we had to convert it to IHEX so
it will not be ignored with Git and be converted back to *.bin,
as usual we'd find in Samsung source.

But specifically in this source, seems like Samsung forgot to add back
IHEX conversion rules. Re-add it from older Kernel e.g. 3.18,
so any *.ihex will be converted back to *.bin and copied to output
directory, just as we expected.

Signed-off-by: Hendra Manudinata <hendra@manudinata.me>